Add openai.textcat.suggest recipe
This recipe accepts a negative_bias parameter from 0 -> 1 that decides
if a particular negative example from ChatGPT will be shown in the
stream.

When filtering, a random number is generated [0, 1]. If that random
number is less than the bias parameter, then that example will be shown.
Else, it is ignored.

Positive examples are always shown in order to correct false positives.