Change how we interact with OpenAISuggester
Instead of inheriting OpenAISuggester and implementing your own
self.parse_response function, we just ask the user to pass a
Callable[str] -> Dict that takes in whatever the raw response of OpenAI
ChatGPT and returns something that's parse-able for Prodigy's view_id.

This should be more in line with how we usually design things.