Add render_vars option for arbitrary variables
Before, I opted for partial rendering in order to include some arbitrary
variables in the Jinja2 template. But I figured that this design is very
shaky and makes things hard to use for another developer. Instead, I
just created a variable, render_vars, for adding arbitrary variables in
the template.